requirements.txt
google-generativeai
pymupdf
pytesseract
Pillow
python-dotenv
